--- BlossomGateway-Core/src/main/java/blossom/gateway/core/netty/NettyHttpServerHandler.java
@@ -1,31 +1,0 @@
-package blossom.gateway.core.netty;
-
-import blossom.gateway.core.context.HttpRequestWrapper;
-import blossom.gateway.core.netty.processor.NettyProcessor;
-import io.netty.channel.ChannelHandlerContext;
-import io.netty.channel.ChannelInboundHandlerAdapter;
-import io.netty.handler.codec.http.FullHttpRequest;
-
-/**
- * @author linqi
- * @version 1.0.0
- * @description
- */
-
-public class NettyHttpServerHandler extends ChannelInboundHandlerAdapter {
-    private final NettyProcessor nettyProcessor;
-
-    public NettyHttpServerHandler(NettyProcessor nettyProcessor) {
-        this.nettyProcessor = nettyProcessor;
-    }
-
-    @Override
-    public void channelRead(ChannelHandlerContext ctx, Object msg) throws Exception {
-        FullHttpRequest request = (FullHttpRequest) msg;
-        HttpRequestWrapper httpRequestWrapper = new HttpRequestWrapper();
-        httpRequestWrapper.setCtx(ctx);
-        httpRequestWrapper.setRequest(request);
-
-        nettyProcessor.process(httpRequestWrapper);
-    }
-}
